refactor(message-bubble): extract memory-proposal-card and message-attachments
Split components/message-bubble.tsx into focused modules:
- memory-proposal-card.tsx: isMemoryProposalAction, getMemoryProposalHeading,
  MemoryProposalCard, MemorySnapshot
- message-attachments.tsx: AttachmentTile, MessageAttachments,
  AssistantInlineImageAttachments
- message-bubble.tsx retains main MessageBubble, helpers, and streaming UI
