+++ components/message-bubble.tsx
@@ -1,21 +1,29 @@
 "use client";
 
 import React, { useEffect, useRef, useState } from "react";
-import { Brain, Check, ChevronDown, ChevronRight, Copy, FileText, GitFork, LoaderCircle, Pencil, RefreshCw, Square, X } from "lucide-react";
+import { Brain, Check, ChevronDown, ChevronRight, Copy, GitFork, LoaderCircle, Pencil, RefreshCw, Square, X } from "lucide-react";
 import { Streamdown } from "streamdown";
 import { code } from "@streamdown/code";
 import { mermaid } from "@streamdown/mermaid";
 import { MarkdownErrorBoundary } from "@/components/markdown-error-boundary";
 import {
   AttachmentPreviewModal,
-  useAttachmentUrlBuilder,
   useAttachmentPreviewController
 } from "@/components/attachment-preview-modal";
 import { CompactionIndicator } from "@/components/compaction-indicator";
-import { Textarea } from "@/components/ui/textarea";
 import { parseAnsiText } from "@/lib/ansi";
 import { stripAttachmentStyleImageMarkdown } from "@/lib/assistant-image-markdown";
 import { writeRichTextToClipboard } from "@/lib/clipboard";
+import {
+  isMemoryProposalAction,
+  getMemoryProposalHeading,
+  MemoryProposalCard
+} from "@/components/memory-proposal-card";
+import {
+  AttachmentTile,
+  MessageAttachments,
+  AssistantInlineImageAttachments
+} from "@/components/message-attachments";
 import type {
   MemoryCategory,
   Message as MessageType,
@@ -24,6 +32,7 @@ import type {
   MessageTimelineItem
 } from "@/lib/types";
 import { normalizeLineBreaks } from "@/lib/text-utils";
+import { Textarea } from "@/components/ui/textarea";
 import {
   Message,
   MessageContent,
@@ -199,316 +208,6 @@ function CollapsibleActionRow({
   );
 }
 
-function isMemoryProposalAction(
-  action: Extract<MessageTimelineItem, { timelineKind: "action" }>
-) {
-  return (
-    (action.kind === "create_memory" ||
-      action.kind === "update_memory" ||
-      action.kind === "delete_memory") &&
-    action.proposalPayload
-  );
-}
-
-function getMemoryProposalHeading(
-  action: Extract<MessageTimelineItem, { timelineKind: "action" }>
-) {
-  const operation = action.proposalPayload?.operation;
-
-  if (action.status === "error") {
-    return "Memory not saved";
-  }
-
-  if (action.proposalState === "approved") {
-    if (operation === "update") {
-      return "Memory updated";
-    }
-
-    if (operation === "delete") {
-      return "Memory deleted";
-    }
-
-    return "Memory saved";
-  }
-
-  if (action.proposalState === "dismissed") {
-    return "Memory ignored";
-  }
-
-  if (action.proposalState === "superseded") {
-    return "Memory proposal superseded";
-  }
-
-  if (operation === "update") {
-    return "Update memory";
-  }
-
-  if (operation === "delete") {
-    return "Delete memory";
-  }
-
-  return "Save memory";
-}
-
-function MemorySnapshot({
-  label,
-  content,
-  category
-}: {
-  label: string;
-  content: string;
-  category: MemoryCategory;
-}) {
-  return (
-    <div className="rounded-md border border-white/6 bg-black/20 px-3 py-2">
-      <p className="text-[10px] font-medium tracking-[0.12em] text-white/45 uppercase">{label}</p>
-      <p className="mt-1 whitespace-pre-wrap break-words text-[12px] leading-5 text-white/84">
-        {content}
-      </p>
-      <p className="mt-1 text-[11px] text-white/48">{category}</p>
-    </div>
-  );
-}
-
-function MemoryProposalCard({
-  action,
-  onApprove,
-  onDismiss,
-  readOnly = false
-}: {
-  action: Extract<MessageTimelineItem, { timelineKind: "action" }>;
-  onApprove?: (
-    actionId: string,
-    overrides?: { content?: string; category?: MemoryCategory }
-  ) => Promise<void>;
-  onDismiss?: (actionId: string) => Promise<void>;
-  readOnly?: boolean;
-}) {
-  const proposal = action.proposalPayload!;
-  const editableMemory = proposal.proposedMemory;
-  const displayMemory = proposal.proposedMemory ?? proposal.currentMemory ?? null;
-  const currentMemory = proposal.currentMemory ?? null;
-  const canEdit = action.kind !== "delete_memory" && Boolean(editableMemory);
-  const isPending = !readOnly && action.status === "pending" && action.proposalState === "pending";
-  const isDelete = proposal.operation === "delete";
-  const heading = getMemoryProposalHeading(action);
-  const [isEditing, setIsEditing] = useState(false);
-  const [draftContent, setDraftContent] = useState(editableMemory?.content ?? "");
-  const [draftCategory, setDraftCategory] = useState<MemoryCategory>(
-    editableMemory?.category ?? currentMemory?.category ?? "other"
-  );
-  const [submissionState, setSubmissionState] = useState<"approve" | "dismiss" | null>(null);
-  const [localError, setLocalError] = useState("");
-
-  useEffect(() => {
-    setDraftContent(editableMemory?.content ?? "");
-    setDraftCategory(editableMemory?.category ?? currentMemory?.category ?? "other");
-    setIsEditing(false);
-    setSubmissionState(null);
-    setLocalError("");
-  }, [action.id, editableMemory?.content, editableMemory?.category, currentMemory?.category]);
-
-  function resetDraft() {
-    setDraftContent(editableMemory?.content ?? "");
-    setDraftCategory(editableMemory?.category ?? currentMemory?.category ?? "other");
-  }
-
-  function handleCancelEdit() {
-    resetDraft();
-    setIsEditing(false);
-    setLocalError("");
-  }
-
-  async function handleApprove() {
-    if (!onApprove) {
-      return;
-    }
-
-    const trimmedContent = draftContent.trim();
-    const nextOverrides =
-      canEdit && !isDelete && trimmedContent
-        ? {
-            content: trimmedContent,
-            category: draftCategory
-          }
-        : undefined;
-
-    setSubmissionState("approve");
-    setLocalError("");
-
-    try {
-      await onApprove(action.id, nextOverrides);
-    } catch (caughtError) {
-      setLocalError(
-        caughtError instanceof Error
-          ? caughtError.message
-          : "Unable to save memory proposal"
-      );
-    } finally {
-      setSubmissionState(null);
-    }
-  }
-
-  async function handleDismiss() {
-    if (!onDismiss) {
-      return;
-    }
-
-    setSubmissionState("dismiss");
-    setLocalError("");
-
-    try {
-      await onDismiss(action.id);
-    } catch (caughtError) {
-      setLocalError(
-        caughtError instanceof Error
-          ? caughtError.message
-          : "Unable to ignore memory proposal"
-      );
-    } finally {
-      setSubmissionState(null);
-    }
-  }
-
-  return (
-    <div className="rounded-lg border border-white/6 bg-white/[0.02] px-3 py-2.5">
-      <div className="flex items-center gap-1.5">
-        <span className="flex h-4 w-4 shrink-0 items-center justify-center rounded-full border border-white/8 bg-white/[0.03]">
-          <Brain className="h-3 w-3 text-violet-400" />
-        </span>
-        <span className="text-[12px] font-medium text-white/88">{heading}</span>
-      </div>
-
-      <div className="mt-2 space-y-2 text-[12px] leading-5 text-white/70">
-        {isEditing && canEdit && isPending ? (
-          <>
-            <Textarea
-              value={draftContent}
-              onChange={(event) => setDraftContent(event.target.value)}
-              className="min-h-[80px] border-white/8 bg-black/20 px-3 py-2 text-[12px] leading-5 text-white focus-visible:ring-0"
-            />
-            <select
-              value={draftCategory}
-              onChange={(event) => setDraftCategory(event.target.value as MemoryCategory)}
-              className="h-9 rounded-md border border-white/8 bg-black/20 px-2.5 text-[12px] text-white outline-none transition focus:border-white/15"
-            >
-              <option value="personal">personal</option>
-              <option value="preference">preference</option>
-              <option value="work">work</option>
-              <option value="location">location</option>
-              <option value="other">other</option>
-            </select>
-          </>
-        ) : proposal.operation === "update" && currentMemory && displayMemory ? (
-          <div className="grid gap-2">
-            <MemorySnapshot
-              label="Before"
-              content={currentMemory.content}
-              category={currentMemory.category}
-            />
-            <MemorySnapshot
-              label="After"
-              content={displayMemory.content}
-              category={displayMemory.category}
-            />
-          </div>
-        ) : proposal.operation === "delete" && currentMemory ? (
-          <>
-            <p className="text-[12px] text-white/68">Remove this memory from saved context.</p>
-            <MemorySnapshot
-              label="Current memory"
-              content={currentMemory.content}
-              category={currentMemory.category}
-            />
-          </>
-        ) : displayMemory ? (
-          <MemorySnapshot
-            label={isPending ? "Proposed memory" : "Memory"}
-            content={displayMemory.content}
-            category={displayMemory.category}
-          />
-        ) : null}
-
-        {action.status === "error" && action.resultSummary ? (
-          <p className="text-[11px] text-red-300">{action.resultSummary}</p>
-        ) : null}
-
-        {localError ? <p className="text-[11px] text-red-300">{localError}</p> : null}
-      </div>
-
-      {isPending ? (
-        <div className="mt-3 flex flex-wrap items-center gap-2">
-          {isEditing && canEdit ? (
-            <>
-              <button
-                type="button"
-                onClick={() => void handleApprove()}
-                disabled={submissionState !== null || !draftContent.trim()}
-                className="inline-flex h-8 items-center justify-center rounded-md border border-white/10 bg-white/[0.06] px-3 text-[12px] font-medium text-white transition hover:bg-white/[0.1] disabled:cursor-not-allowed disabled:opacity-50"
-              >
-                {submissionState === "approve" ? "Saving..." : "Save"}
-              </button>
-              <button
-                type="button"
-                onClick={handleCancelEdit}
-                disabled={submissionState !== null}
-                className="inline-flex h-8 items-center justify-center rounded-md border border-white/8 bg-transparent px-3 text-[12px] font-medium text-white/72 transition hover:border-white/14 hover:text-white disabled:cursor-not-allowed disabled:opacity-50"
-              >
-                Cancel
-              </button>
-            </>
-          ) : (
-            <>
-              <button
-                type="button"
-                onClick={() => void handleApprove()}
-                disabled={submissionState !== null}
-                className="inline-flex h-8 items-center justify-center rounded-md border border-white/10 bg-white/[0.06] px-3 text-[12px] font-medium text-white transition hover:bg-white/[0.1] disabled:cursor-not-allowed disabled:opacity-50"
-              >
-                {submissionState === "approve"
-                  ? isDelete
-                    ? "Deleting..."
-                    : "Saving..."
-                  : isDelete
-                    ? "Delete memory"
-                    : "Save"}
-              </button>
-              <button
-                type="button"
-                onClick={() => void handleDismiss()}
-                disabled={submissionState !== null}
-                className="inline-flex h-8 items-center justify-center rounded-md border border-white/8 bg-transparent px-3 text-[12px] font-medium text-white/72 transition hover:border-white/14 hover:text-white disabled:cursor-not-allowed disabled:opacity-50"
-              >
-                {submissionState === "dismiss"
-                  ? isDelete
-                    ? "Cancelling..."
-                    : "Ignoring..."
-                  : isDelete
-                    ? "Cancel"
-                    : "Ignore"}
-              </button>
-              {canEdit ? (
-                <button
-                  type="button"
-                  onClick={() => {
-                    resetDraft();
-                    setIsEditing(true);
-                    setLocalError("");
-                  }}
-                  disabled={submissionState !== null}
-                  className="inline-flex h-8 items-center justify-center rounded-md border border-white/8 bg-transparent px-3 text-[12px] font-medium text-white/72 transition hover:border-white/14 hover:text-white disabled:cursor-not-allowed disabled:opacity-50"
-                >
-                  Edit
-                </button>
-              ) : null}
-            </>
-          )}
-        </div>
-      ) : null}
-    </div>
-  );
-}
-
 const ASSISTANT_MAX_WIDTH = "max-w-full md:max-w-[95%]";
 const ASSISTANT_BUBBLE =
   "w-fit max-w-full rounded-2xl border border-white/8 bg-white/[0.03] px-2.5 py-2 md:px-4 md:py-3 text-[var(--text)] shadow-[0_8px_24px_rgba(0,0,0,0.28)]";
@@ -533,136 +232,6 @@ function formatPlainTextHtml(value: string) {
     .split(/\n{2,}/)
     .map((paragraph) => `<p>${escapeHtml(paragraph).replace(/\n/g, "<br />")}</p>`)
     .join("");
-}
-
-function AttachmentTile({
-  attachment,
-  compact = false,
-  onPreview
-}: {
-  attachment: MessageAttachment;
-  compact?: boolean;
-  onPreview: (attachment: MessageAttachment) => void;
-}) {
-  const buildAttachmentUrl = useAttachmentUrlBuilder();
-
-  if (attachment.kind === "image") {
-    return (
-      <button
-        type="button"
-        aria-label={`Preview ${attachment.filename}`}
-        onClick={() => onPreview(attachment)}
-        className={`overflow-hidden rounded-xl border border-white/10 bg-black/20 ${compact ? "w-16" : "w-28"}`}
-      >
-        {/* eslint-disable-next-line @next/next/no-img-element -- Attachment thumbnails are API-served user files that next/image cannot safely optimize. */}
-        <img
-          src={buildAttachmentUrl(attachment)}
-          alt=""
-          aria-hidden="true"
-          className={`w-full object-cover ${compact ? "h-16" : "h-28"}`}
-        />
-      </button>
-    );
-  }
-
-  return (
-    <button
-      type="button"
-      aria-label={`Preview ${attachment.filename}`}
-      onClick={() => onPreview(attachment)}
-      className="flex min-w-0 items-center gap-2 rounded-xl border border-white/10 bg-black/15 px-3 py-2 text-left"
-    >
-      <span className="flex h-8 w-8 shrink-0 items-center justify-center rounded-lg bg-white/10 text-white/75">
-        <FileText className="h-4 w-4" />
-      </span>
-      <span className="min-w-0">
-        <span className="block truncate text-sm font-medium text-white">{attachment.filename}</span>
-        <span className="block truncate text-xs text-white/60">{attachment.mimeType}</span>
-      </span>
-    </button>
-  );
-}
-
-function MessageAttachments({
-  attachments,
-  compact = false,
-  onPreview
-}: {
-  attachments: MessageAttachment[];
-  compact?: boolean;
-  onPreview: (attachment: MessageAttachment) => void;
-}) {
-  if (!attachments.length) {
-    return null;
-  }
-
-  const images = attachments.filter((attachment) => attachment.kind === "image");
-  const files = attachments.filter((attachment) => attachment.kind === "text");
-
-  return (
-    <div className="space-y-2.5">
-      {images.length ? (
-        <div className="flex flex-wrap gap-2">
-          {images.map((attachment) => (
-            <AttachmentTile
-              key={attachment.id}
-              attachment={attachment}
-              compact={compact}
-              onPreview={onPreview}
-            />
-          ))}
-        </div>
-      ) : null}
-      {files.length ? (
-        <div className="space-y-2">
-          {files.map((attachment) => (
-            <AttachmentTile
-              key={attachment.id}
-              attachment={attachment}
-              compact={compact}
-              onPreview={onPreview}
-            />
-          ))}
-        </div>
-      ) : null}
-    </div>
-  );
-}
-
-function AssistantInlineImageAttachments({
-  attachments,
-  onPreview
-}: {
-  attachments: MessageAttachment[];
-  onPreview: (attachment: MessageAttachment) => void;
-}) {
-  const buildAttachmentUrl = useAttachmentUrlBuilder();
-
-  if (!attachments.length) {
-    return null;
-  }
-
-  return (
-    <div className="flex flex-wrap gap-2.5">
-      {attachments.map((attachment) => (
-        <button
-          key={attachment.id}
-          type="button"
-          aria-label={`Preview ${attachment.filename}`}
-          onClick={() => onPreview(attachment)}
-          className="max-w-full overflow-hidden rounded-xl border border-white/10 bg-black/20"
-        >
-          {/* eslint-disable-next-line @next/next/no-img-element -- Attachment previews are API-served user files that next/image cannot safely optimize. */}
-          <img
-            src={buildAttachmentUrl(attachment)}
-            alt=""
-            aria-hidden="true"
-            className="block max-h-[28rem] w-auto max-w-full object-contain"
-          />
-        </button>
-      ))}
-    </div>
-  );
 }
 
 export function MessageBubble({
